requirements.txt
Flask
Flask-SQLAlchemy
langchain_experimental
langchain
openai
python-dotenv